% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.yadir_token}
\alias{print.yadir_token}
\title{Print yadir_token}
\usage{
\method{print}{yadir_token}(x, ...)
}
\arguments{
\item{x}{yadir_token object}

\item{...}{extra arguments}
}
\description{
S3 method for printing yadir_token objects
}
