\name{ryandexdirect-deprecated}
\alias{ryandexdirect-deprecated}
\title{Deprecated functions in ryandexdirect}
\description{List of depricated functions and their new versions.}
\section{Deprecated fnctions}{
Next function is deprecated:
\describe{
  \item{\code{yadirGetCampaignList()}}{use \code{\link[=yadirGetCampaign]{yadirGetCampaign()}}, from september 2019}
  \item{yadirGetCampaignListOld()}{use \code{\link[=yadirGetCampaign]{yadirGetCampaign()}}, from april 2019}
  \item{yadirGetLogsData()}{use \code{\link[rym:rym_get_logs]{rym_get_logs()}}, from april 2019}
  \item{yadirGetMetrikaGAData()}{use \code{\link[rym:rym_get_ga]{rym_get_ga()}}, from april 2019}
  \item{yadirGetSummaryStat()}{use \code{\link[=yadirGetReport]{yadirGetReport()}}, from march 2017}
}

}