\name{ryandexdirect-package}
\alias{ryandexdirect-package}
\alias{ryandexdirect}
\docType{package}
\title{
\packageTitle{ryandexdirect}
}
\description{
\packageDescription{ryandexdirect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ryandexdirect}
\packageIndices{ryandexdirect}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ryandexdirect}

Maintainer: \packageMaintainer{ryandexdirect}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
